pyproject.toml
[tool.poetry]
name = "recommend-movie"
version = "0.1.0"
description = ""
authors = ["ac2393921 <ac2393921@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.0"
isort = "^5.12.0"
pydantic = "^1.10.7"
pandas = "^2.0.0"
pandera = "^0.14.5"
mlxtend = "^0.22.0"
faiss-cpu = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
